feat: Unify authentication methods (#58)
* unify iam and app auth
* update app list api url
* update strategy for selecting app
* rename auth
* add access_key cache
* update readme
* fix ut
* update _check_error & rename qianfan_requestor
