+++ src/qianfan/resources/auth/oauth.py
@@ -14,15 +14,12 @@
 
 import threading
 import time
-from typing import Any, Dict, Optional, Tuple
-
-from baidubce.auth.bce_credentials import BceCredentials
-from baidubce.auth.bce_v1_signer import sign
-from baidubce.utils import get_canonical_time
+from typing import Any, Dict, List, Optional, Tuple
 
 from qianfan.config import get_config
 from qianfan.consts import Consts
 from qianfan.errors import InternalError, InvalidArgumentError
+from qianfan.resources.console.app import _App
 from qianfan.resources.http_client import HTTPClient
 from qianfan.resources.typing import QfRequest, RetryConfig
 from qianfan.utils import (
@@ -255,7 +252,15 @@ class Auth(object):
     _ak: Optional[str] = None
     _sk: Optional[str] = None
     _access_token: Optional[str] = None
+    _access_key: Optional[str] = None
+    _secret_key: Optional[str] = None
     _registered: bool = False
+    _console_ak_to_app_ak: Dict[Tuple[str, str], Tuple[str, str]] = {}
+    """
+    (access_key, secret_key) -> (ak, sk)
+    map which convert console ak/sk to qianfan ak/sk
+    use as cache to avoid querying console ak/sk multple times
+    """
 
     def __init__(self, **kwargs: Any) -> None:
         """
@@ -271,12 +276,77 @@ class Auth(object):
         self._access_token = (
             kwargs.get("access_token", None) or get_config().ACCESS_TOKEN
         )
+        self._access_key = kwargs.get("access_key", None) or get_config().ACCESS_KEY
+        self._secret_key = kwargs.get("secret_key", None) or get_config().SECRET_KEY
+        if not self._credential_available():
+            raise InvalidArgumentError(
+                "no enough credential found, any one of (access_key, secret_key),"
+                " (ak, sk), access_token must be provided"
+            )
+
+    def _update_ak_sk_from_app_list(self, app_list: List[Dict[str, Any]]) -> None:
+        """
+        select the latest app from `app_list` and update ak, sk
+        """
+        if len(app_list) == 0:
+            raise InvalidArgumentError(
+                "no app found, please create an app first on qianfan console:"
+                " https://console.bce.baidu.com/qianfan/ais/console/applicationConsole/application"
+            )
+        appid = get_config().APPID
+        # the last app is the latest one
+        selected_app = app_list[-1]
+        if appid is None:
+            # if user does not provide appid, use the latest one
+            log_info(
+                "no appid provided, using the latest one which id is"
+                f" {selected_app['id']}"
+            )
+        else:
+            # find the uesr required app
+            tmp = [app for app in app_list if app["id"] == appid]
+            if len(tmp) == 0:
+                log_warn(
+                    f"The provided appid {appid} is not found, please check the"
+                    " id and the sdk will use the latest one which id is"
+                    f" {selected_app['id']}"
+                )
+            else:
+                selected_app = tmp[0]
+        self._ak = selected_app["ak"]
+        self._sk = selected_app["sk"]
 
     def _register(self) -> None:
         """
         register the access token to manager, so that it can be refreshed automatically
         """
         if not self._registered:
+            # when user only provides access_key and secret_key
+            # sdk needs to get ak and sk by access_key
+            if (
+                self._access_key is not None
+                and self._secret_key is not None
+                and self._ak is None
+                and self._sk is None
+            ):
+                self._ak, self._sk = Auth._console_ak_to_app_ak.get(
+                    (self._access_key, self._secret_key), (None, None)
+                )
+                if self._ak is None:
+                    resp = _App.list(
+                        access_key=self._access_key, secret_key=self._secret_key
+                    )
+                    app_list = resp["result"]["appList"]
+                    self._update_ak_sk_from_app_list(app_list)
+                    if self._ak is None or self._sk is None:
+                        # ak and sk should already be set
+                        # otherwise an exception should already be raised
+                        raise InternalError
+                    Auth._console_ak_to_app_ak[(self._access_key, self._secret_key)] = (
+                        self._ak,
+                        self._sk,
+                    )
+
             if self._access_token is None:
                 # if access_token is not provided, both ak and sk should be provided
                 if self._ak is None or self._sk is None:
@@ -297,6 +367,18 @@ class Auth(object):
         register the access token to manager, so that it can be refreshed automatically
         """
         if not self._registered:
+            if (
+                self._access_key is not None
+                and self._secret_key is not None
+                and self._ak is None
+                and self._sk is None
+            ):
+                resp = await _App.alist(
+                    access_key=self._access_key, secret_key=self._secret_key
+                )
+                app_list = resp["result"]["appList"]
+                self._update_ak_sk_from_app_list(app_list)
+
             if self._access_token is None:
                 # if access_token is not provided, both ak and sk should be provided
                 if self._ak is None or self._sk is None:
@@ -323,6 +405,7 @@ class Auth(object):
             return
         self._register()
         AuthManager().refresh_access_token(self._ak, self._sk)
+        self._access_token = None
 
     async def arefresh_access_token(self) -> None:
         """
@@ -333,53 +416,41 @@ class Auth(object):
             return
         await self._aregister()
         await AuthManager().arefresh_access_token(self._ak, self._sk)
+        self._access_token = None
+
+    def _credential_available(self) -> bool:
+        if self._access_token is not None:
+            return True
+        if self._ak is not None and self._sk is not None:
+            return True
+        if self._access_key is not None and self._secret_key is not None:
+            return True
+        return False
 
     def access_token(self) -> str:
         """
         get current `access_token`
         """
-        if self._ak is None or self._sk is None:
-            if self._access_token is None:
-                log_warn("Access token is not available! Please check code!")
-                return ""
+        if self._access_token is not None and (self._ak is None or self._sk is None):
             return self._access_token
         self._register()
+        if self._ak is None or self._sk is None:
+            raise InvalidArgumentError(
+                "no enough credential found, any one of (access_key, secret_key),"
+                " (ak, sk), access_token must be provided"
+            )
         return AuthManager().get_access_token(self._ak, self._sk)
 
     async def a_access_token(self) -> str:
         """
         get current `access_token`
         """
-        if self._ak is None or self._sk is None:
-            if self._access_token is None:
-                log_warn("Access token is not available! Please check code!")
-                return ""
+        if self._access_token is not None and (self._ak is None or self._sk is None):
             return self._access_token
         await self._aregister()
+        if self._ak is None or self._sk is None:
+            raise InvalidArgumentError(
+                "no enough credential found, any one of (access_key, secret_key),"
+                " (ak, sk), access_token must be provided"
+            )
         return await AuthManager().aget_access_token(self._ak, self._sk)
-
-
-def iam_sign(
-    ak: str,
-    sk: str,
-    request: QfRequest,
-) -> None:
-    """
-    Create the authorization
-    """
-    credentials = BceCredentials(ak, sk)
-    timestamp = time.time()
-    x_bce_date = get_canonical_time(timestamp)
-    request.headers["x-bce-date"] = x_bce_date
-    authorization = sign(
-        credentials,
-        str.encode(request.method),
-        str.encode(request.url),
-        {str.encode(k): v for k, v in request.headers.items()},
-        {str.encode(k): v for k, v in request.query.items()},
-        timestamp=timestamp,
-        expiration_in_seconds=get_config().IAM_SIGN_EXPIRATION_SEC,
-        headers_to_sign={str.encode(k.lower()) for k in request.headers.keys()},
-    )
-
-    request.headers["Authorization"] = authorization
